requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
autopep8==1.4.4
Django==2.2
django-filter==2.1.0
django-templated-mail==1.1.1
djangorestframework==3.9.2
djoser==1.5.1
model-mommy==1.6.0
more-itertools==7.0.0
mysqlclient==1.4.2.post1
Pillow==5.4.1
pkg-resources==0.0.0
pluggy==0.11.0
py==1.8.0
pycodestyle==2.5.0
pytest==4.4.2
pytest-django==3.4.8
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
